+++ src/index.ts
@@ -1,4 +1,5 @@
-export { LinkIO } from './LinkIO';
-export { InMemoryStorage } from './storage/InMemoryStorage';
-export { RedisStorage } from './storage/RedisStorage';
-export * from './types';
+export { LinkIO } from "./LinkIO";
+export { InMemoryStorage } from "./storage/InMemoryStorage";
+export { RedisStorage } from "./storage/RedisStorage";
+export { generateFingerprint, getClientIP } from "./utils";
+export * from "./types";